RR-721 - Added type as per PR suggestion
